# api-docs.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "FAM Backend API",
    "version": "1.0.0",
    "description": "Family event management platform — three-role auth (user/admin/organizer), family tree, QR join, notifications."
  },
  "servers": [
    {
      "url": "http://localhost:8000"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "EventObject": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "example": "64b1f2c3e4d5f6a7b8c9d0e1" },
          "name": { "type": "string", "example": "Smith Wedding" },
          "code": { "type": "string", "example": "smithweddi4821" },
          "qrCode": { "type": "string", "example": "https://res.cloudinary.com/demo/image/upload/v1/qr/event_64b1f2c3.png" },
          "eventDate": { "type": "string", "format": "date-time" },
          "approvalMode": { "type": "string", "enum": ["auto", "manual"] }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/admin/login": {
      "post": {
        "summary": "Admin login with name + password",
        "description": "Rate limited: 5 req/15 min. Login uses name, not email.",
        "tags": ["Admin"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name", "password"],
                "properties": {
                  "name": { "type": "string", "example": "admin" },
                  "password": { "type": "string", "example": "secret123" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Returns access token (24h) and admin object",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "message": { "type": "string", "example": "Admin logged in successfully" },
                    "token": { "type": "string" },
                    "admin": {
                      "type": "object",
                      "properties": {
                        "id": { "type": "string" },
                        "name": { "type": "string" },
                        "email": { "type": "string" },
                        "role": { "type": "string", "example": "admin" }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": { "description": "Name and password are required" },
          "401": { "description": "Invalid credentials" }
        }
      }
    },
    "/api/admin/logout": {
      "post": {
        "summary": "Admin logout",
        "tags": ["Admin"],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Logged out",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "message": { "type": "string", "example": "Admin logged out successfully" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/profile": {
      "get": {
        "summary": "Get admin profile (no password field)",
        "tags": ["Admin"],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Admin profile",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "admin": { "type": "object" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/dashboard/stats": {
      "get": {
        "summary": "Get admin dashboard statistics",
        "tags": ["Admin"],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Platform-wide counts",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "stats": {
                      "type": "object",
                      "properties": {
                        "totalUsers": { "type": "integer", "example": 120 },
                        "totalOrganizers": { "type": "integer", "example": 8 },
                        "totalEvents": { "type": "integer", "example": 5 },
                        "activeEvents": { "type": "integer", "example": 3 },
                        "blockedUsers": { "type": "integer", "example": 2 },
                        "totalRelationships": { "type": "integer", "example": 340 },
                        "newUsersLast7Days": { "type": "integer", "example": 15 },
                        "newEventsLast7Days": { "type": "integer", "example": 1 }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/events": {
      "post": {
        "summary": "Create a new event — automatically creates an Organizer account (temp password emailed)",
        "tags": ["Admin"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["eventName", "eventDate", "organizerEmail", "organizerName"],
                "properties": {
                  "eventName": { "type": "string", "example": "Smith Wedding" },
                  "organizerEmail": { "type": "string", "example": "org@example.com" },
                  "organizerName": { "type": "string", "example": "Sarah Jones" },
                  "description": { "type": "string", "example": "Annual family wedding celebration" },
                  "eventType": {
                    "type": "string",
                    "enum": ["wedding", "reunion", "birthday", "funeral", "festival", "other"],
                    "default": "other"
                  },
                  "eventDate": { "type": "string", "format": "date", "example": "2026-06-15" },
                  "eventEndDate": { "type": "string", "format": "date", "example": "2026-06-16" },
                  "approvalMode": { "type": "string", "enum": ["auto", "manual"], "default": "manual" },
                  "settings": {
                    "type": "object",
                    "properties": {
                      "approvalRequired": { "type": "boolean" },
                      "maxMembersPerUser": { "type": "integer", "example": 4 }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Event and organizer created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "message": { "type": "string", "example": "Event created successfully" },
                    "event": { "$ref": "#/components/schemas/EventObject" },
                    "organizer": {
                      "type": "object",
                      "properties": {
                        "id": { "type": "string" },
                        "name": { "type": "string" },
                        "email": { "type": "string" },
                        "tempPassword": { "type": "string", "example": "abc12345" },
                        "accessCode": { "type": "string", "example": "XY12AB" }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": { "description": "Missing required fields" }
        }
      },
      "get": {
        "summary": "Paginated list of all events",
        "tags": ["Admin"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "in": "query", "name": "page", "schema": { "type": "integer", "default": 1 } },
          { "in": "query", "name": "limit", "schema": { "type": "integer", "default": 20 } },
          { "in": "query", "name": "isActive", "schema": { "type": "boolean" } }
        ],
        "responses": {
          "200": { "description": "Paginated array of event objects" }
        }
      }
    },
    "/api/admin/events/summary": {
      "get": {
        "summary": "All events with participant and relationship counts (no pagination)",
        "tags": ["Admin"],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "Events with user counts" }
        }
      }
    },
    "/api/admin/events/lookup/{identifier}": {
      "get": {
        "summary": "Get event by MongoDB ID or event code — includes rich stats and recent activity",
        "tags": ["Admin"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "in": "path",
            "name": "identifier",
            "required": true,
            "schema": { "type": "string" },
            "description": "MongoDB ObjectId or event code (e.g. smithweddi4821)"
          }
        ],
        "responses": {
          "200": { "description": "Event with stats, top contributors, recent activity" },
          "404": { "description": "Event not found" }
        }
      }
    },
    "/api/admin/events/id/{id}": {
      "get": {
        "summary": "Get event by MongoDB ObjectId",
        "tags": ["Admin"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "in": "path", "name": "id", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Event object" },
          "404": { "description": "Not found" }
        }
      }
    },
    "/api/admin/events/code/{code}": {
      "get": {
        "summary": "Get event by event code",
        "tags": ["Admin"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "in": "path", "name": "code", "required": true, "schema": { "type": "string" }, "example": "smithweddi4821" }
        ],
        "responses": {
          "200": { "description": "Event object" },
          "404": { "description": "Not found" }
        }
      }
    },
    "/api/admin/events/{id}": {
      "delete": {
        "summary": "Soft-delete an event (sets isActive=false). Also deactivates all organizers for the event.",
        "tags": ["Admin"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "in": "path", "name": "id", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Event deleted" },
          "404": { "description": "Event not found" }
        }
      }
    },
    "/api/admin/events/{id}/approval-mode": {
      "patch": {
        "summary": "Update event join approval mode",
        "tags": ["Admin"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "in": "path", "name": "id", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["approvalMode"],
                "properties": {
                  "approvalMode": { "type": "string", "enum": ["auto", "manual"] }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Approval mode updated" }
        }
      }
    },
    "/api/admin/events/{id}/tree-visibility": {
      "patch": {
        "summary": "Update tree visibility for an event",
        "tags": ["Admin"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "in": "path", "name": "id", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "treeVisibility": {
                    "type": "string",
                    "enum": ["participants", "organizer_only", "admin_only"]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Tree visibility updated" }
        }
      }
    },
    "/api/admin/events/{eventId}/users": {
      "get": {
        "summary": "All participants, pending approvals, organizers and top contributors for a specific event",
        "tags": ["Admin"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "in": "path", "name": "eventId", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Participants, pending approvals, organizers, top contributors" }
        }
      }
    },
    "/api/admin/events/{eventId}/users/{userId}": {
      "get": {
        "summary": "Detailed profile, relationships and stats for one user within an event",
        "tags": ["Admin"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "in": "path", "name": "eventId", "required": true, "schema": { "type": "string" } },
          { "in": "path", "name": "userId", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "User event detail with relationships and stats" }
        }
      }
    },
    "/api/admin/organizers": {
      "post": {
        "summary": "Create a standalone organizer (not tied to event creation)",
        "description": "Only name and email are required. assignedEvent and validTill are optional (validTill defaults to 90 days from now). Default permissions: moderate_tree, approve_users, manage_schedule, view_participants.",
        "tags": ["Admin"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name", "email"],
                "properties": {
                  "name": { "type": "string", "example": "Mike Chen" },
                  "email": { "type": "string", "example": "mike@example.com" },
                  "assignedEvent": { "type": "string", "description": "MongoDB ObjectId of the event (optional)" },
                  "validTill": { "type": "string", "format": "date", "example": "2026-12-31" }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Organizer created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "message": { "type": "string", "example": "Organizer created successfully" },
                    "organizer": {
                      "type": "object",
                      "properties": {
                        "id": { "type": "string" },
                        "name": { "type": "string" },
                        "email": { "type": "string" },
                        "accessCode": { "type": "string", "example": "ABCD1234" },
                        "tempPassword": { "type": "string", "example": "xyz9876543" },
                        "assignedEvent": { "type": "string", "nullable": true },
                        "validTill": { "type": "string", "format": "date-time" }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": { "description": "Name and email are required, or email already exists" }
        }
      },
      "get": {
        "summary": "All organizers with populated assignedEvent and createdBy",
        "tags": ["Admin"],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Array of organizer objects",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "count": { "type": "integer" },
                    "organizers": { "type": "array", "items": { "type": "object" } }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/organizers/{id}": {
      "get": {
        "summary": "Single organizer with populated event and creator",
        "tags": ["Admin"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "in": "path", "name": "id", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Organizer object" },
          "404": { "description": "Not found" }
        }
      },
      "patch": {
        "summary": "Update organizer fields",
        "tags": ["Admin"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "in": "path", "name": "id", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string" },
                  "email": { "type": "string" },
                  "assignedEvent": { "type": "string" },
                  "validTill": { "type": "string", "format": "date" },
                  "isActive": { "type": "boolean" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Organizer updated successfully" },
          "404": { "description": "Not found" }
        }
      },
      "delete": {
        "summary": "Soft-delete organizer (sets isActive=false)",
        "tags": ["Admin"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "in": "path", "name": "id", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Organizer removed successfully" },
          "404": { "description": "Not found" }
        }
      }
    },
    "/api/admin/organizers/{id}/reactivate": {
      "patch": {
        "summary": "Reactivate a deactivated organizer (sets isActive=true)",
        "tags": ["Admin"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "in": "path", "name": "id", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Organizer reactivated" },
          "404": { "description": "Not found" }
        }
      }
    },
    "/api/admin/users": {
      "get": {
        "summary": "Paginated list of all users",
        "tags": ["Admin"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "in": "query", "name": "page", "schema": { "type": "integer", "default": 1 } },
          { "in": "query", "name": "limit", "schema": { "type": "integer", "default": 50 } }
        ],
        "responses": {
          "200": { "description": "Array of user objects" }
        }
      }
    },
...
  },
...
}